separated ui and model_checks, constants, ui_check_base
