+++ .gitignore
@@ -145,6 +145,9 @@ venv.bak/
 .dmypy.json
 dmypy.json
 
+# pylint
+.pylintrc
+
 # Pyre type checker
 .pyre/
